updated logic (login protection is opposite of reg protection so in need of its own function) + moved the redirect outside of commons folder in libs/mw
