+++ src/lib/middleware/auth/protectedRoute.ts
@@ -1,15 +1,24 @@
 import type { NextRequest } from "next/server";
-import redirectTo from "../common/redirect";
+import redirectTo from "../redirect";
 import { getSessionCookie } from "better-auth/cookies";
 
-const handleProtectedRoutes = async (request: NextRequest, protectedRoutes: any[]) => {
+export const handleProtectedRoutes = async (request: NextRequest, protectedRoutes: any[]) => {
     const sessionCookie = await getSessionCookie(request);
   
-    if (sessionCookie && protectedRoutes.includes(request.nextUrl.pathname)) {
-      return redirectTo(request, '/')
+    if (!sessionCookie && protectedRoutes.includes(request.nextUrl.pathname)) {
+      return redirectTo(request, '/auth/signin')
     }
   
     return null;
 };
+
+// Special case in which logic is slightly different (therefore warrants its own function)
+export const handleSignInRoutes = async (request: NextRequest, protectedRoutes: any[]) => {
+  const sessionCookie = await getSessionCookie(request);
   
-export default handleProtectedRoutes
+  if (sessionCookie && protectedRoutes.includes(request.nextUrl.pathname)) {
+    return redirectTo(request, '/')
+  }
+
+  return null;
+}
