+++ src/middleware.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import handleProtectedRoutes from './lib/middleware/auth/protectedRoute';
+import {handleProtectedRoutes, handleSignInRoutes} from './lib/middleware/auth/protectedRoute';
 import protectedLoginRoutes from './data/protected-routes/protectedLoginRoutes';
 import protectedUserRoutes from './data/protected-routes/protectedUserRoutes';
 
@@ -9,7 +9,7 @@ export const middleware = async (request: NextRequest) => {
   if (userResponse) return userResponse;
   
 
-  const signinResponse = await handleProtectedRoutes(request, protectedLoginRoutes);
+  const signinResponse = await handleSignInRoutes(request, protectedLoginRoutes);
   if (signinResponse) return signinResponse;
